ui-builder/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "site-builder"
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
description = "Site builder "
dependencies = [
  "instaloader",
  "pyyaml",
  "typer",
  "jinja2",
  "pillow",
] # Add project dependencies here, e.g. ["click", "numpy"]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
  "copier",
  "pipdeptree",
  "pre-commit",
  "pyright",
  "pytest",
  "pytest-cov",
  "ruff",
  "tox-direct",
  "types-mock",
]

[project.scripts]
site-builder = "site_builder.__main__:app"

[project.urls]
GitHub = "https://github.com/maxherbs/site-builder"

[[project.authors]] # Further authors may be added by duplicating this section
email = "max@maxherbert.com"
name = "Max Herbert"


[tool.setuptools_scm]
version_file = "src/site_builder/_version.py"

[tool.pyright]
typeCheckingMode = "standard"
reportMissingImports = false  # Ignore missing stubs in imported modules

[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
filterwarnings = "error"
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"

[tool.coverage.run]
data_file = "/tmp/site_builder.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,type-checking,tests}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pytest
    pre-commit
    pyright
commands =
    pre-commit: pre-commit run --all-files --show-diff-on-failure {posargs}
    type-checking: pyright src tests {posargs}
    tests: pytest --cov=site_builder --cov-report term --cov-report xml:cov.xml {posargs}
"""

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.select = [
  "B",   # flake8-bugbear - https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
  "C4",  # flake8-comprehensions - https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
  "E",   # pycodestyle errors - https://docs.astral.sh/ruff/rules/#error-e
  "F",   # pyflakes rules - https://docs.astral.sh/ruff/rules/#pyflakes-f
  "W",   # pycodestyle warnings - https://docs.astral.sh/ruff/rules/#warning-w
  "I",   # isort - https://docs.astral.sh/ruff/rules/#isort-i
  "UP",  # pyupgrade - https://docs.astral.sh/ruff/rules/#pyupgrade-up
  "SLF", # self - https://docs.astral.sh/ruff/settings/#lintflake8-self
]

[tool.ruff.lint.per-file-ignores]
# By default, private member access is allowed in tests
# See https://github.com/DiamondLightSource/python-copier-template/issues/154
# Remove this line to forbid private member access in tests
"tests/**/*" = ["SLF001"]
